[Refactoring] Move command function
