+++ index.js
@@ -13,10 +13,9 @@ const {
   ButtonStyle,
 } = require("discord.js");
 const DB = require("./db_api.js");
+const COMMAND = require("./command.js");
 const { token } = require("./config-dev.json"); //commit 시 수정
 
-let team1Temp = [];
-let team2Temp = [];
 let checkDelay = false;
 
 const client = new Client({
@@ -31,21 +30,6 @@ const client = new Client({
   partials: [Partials.Channel],
 });
 
-let btnRow = new ActionRowBuilder().setComponents(
-  new ButtonBuilder()
-    .setCustomId("team1winBtn")
-    .setLabel("1️⃣팀 승리")
-    .setStyle(ButtonStyle.Primary),
-  new ButtonBuilder()
-    .setCustomId("team2winBtn")
-    .setLabel("2️⃣팀 승리")
-    .setStyle(ButtonStyle.Primary),
-  new ButtonBuilder()
-    .setCustomId("rerollBtn")
-    .setLabel("🎲리롤🎲")
-    .setStyle(ButtonStyle.Danger)
-);
-
 client.once(Events.ClientReady, (c) => {
   console.log(`Ready! Logged in as ${c.user.tag}`);
 });
@@ -54,86 +38,28 @@ client.on("error", (err) => {
   console.log(err.message);
 });
 
-async function registerUser(message) {
-  let onlineUserArr = [];
-  message.guild.members.fetch().then((fetchedMembers) => {
-    fetchedMembers.forEach((k, v) => {
-      let curMem = message.guild.members.cache.get(k.id);
-      if (curMem.voice.channel) {
-        onlineUserArr.push(k.user.username);
-      }
-    });
-    onlineUserArr.forEach(async (item) => {
-      await DB.addToDatabase(item);
-    });
-  });
-}
-
-function makeTeam(message) {
-  let onlineUserArr = [];
-  message.guild.members.fetch().then((fetchedMembers) => {
-    fetchedMembers.forEach((k, v) => {
-      let curMem = message.guild.members.cache.get(k.id);
-      if (curMem.voice.channel) {
-        onlineUserArr.push(k.user.username);
-      }
-    });
-    if (onlineUserArr.length == 0) {
-      ("현재 채널 접속 인원이 없습니다.");
-    } else if (onlineUserArr.length % 2 === 1) {
-      message.guild.channels.cache
-        .get(message.channelId)
-        .send("현재 채널 접속 인원이 홀수입니다. 짝수 인원으로 맞춰주세요.");
-    } else {
-      const teamCount = onlineUserArr.length / 2;
-      const generateRandomKey = () => Math.floor(Math.random() * 1000);
-      const userWithKey = onlineUserArr.map((user) => ({
-        name: user,
-        key: generateRandomKey(),
-      }));
-      const sortedUserWithKey = userWithKey.sort((a, b) => {
-        return a.key >= b.key ? 1 : -1;
-      });
-      const team1 = sortedUserWithKey.splice(0, teamCount);
-      const team2 = sortedUserWithKey.splice(-teamCount);
-
-      team1Temp = [];
-      team2Temp = [];
-      team1.forEach((usr) => {
-        team1Temp.push(usr.name);
-      });
-      team2.forEach((usr) => {
-        team2Temp.push(usr.name);
-      });
-      const exampleEmbed = new EmbedBuilder()
-        .setColor(0x0099ff)
-        .setTitle("팀 구성 결과🚀")
-        .setURL("https://discord.js.org/")
-        .addFields(
-          { name: "1️⃣팀", value: team1Temp.join(", ") },
-          { name: "2️⃣팀", value: team2Temp.join(", ") }
-        );
-      message.channel.send({ embeds: [exampleEmbed], components: [btnRow] });
-    }
-  });
-}
-
 client.on("messageCreate", async (message) => {
   if (message.author.username === "kwonSM") {
     message.react("💩");
   }
-  if (message.content == "!5vs5") {
-    message.reply("@everyone");
+  switch(message.content){
+    case "!5vs5":
+      message.reply("@everyone");
+      break;
+    case "!ping":
+      message.reply("pong");
+      break;
+    case "!dice":
+      message.reply(`🎲${message.author.username}님의 주사위: ${COMMAND.rollDice().toString()}🎲`);
+      break;
+    case "!team":
+      COMMAND.makeTeam(message);
+      setTimeout(() => {
+        checkDelay = true;
+      }, 60000);
+      break;
   }
-  if (message.content == "!ping") {
-    message.reply("pong");
-  }
-  if (message.content == "!dice") {
-    message.channel.send("🎲주사위 굴리는 중...🎲");
-    setTimeout(() => {
-      message.reply(Math.floor(Math.random() * 100).toString());
-    }, 1500);
-  }
+  
   if (message.content == "!help") {
     const explain = new EmbedBuilder()
       .setColor(0x0099ff)
@@ -148,12 +74,6 @@ client.on("messageCreate", async (message) => {
         { name: "!5vs5", value: "5대5 소집령" }
       );
     message.reply({ embeds: [explain] });
-  }
-  if (message.content == "!team") {
-    makeTeam(message);
-    setTimeout(() => {
-      checkDelay = true;
-    }, 60000);
   }
   if (message.content == "!top3") {
     const top3Data = await DB.getTop3(true);
@@ -177,7 +97,7 @@ client.on("messageCreate", async (message) => {
     message.reply({ embeds: [exampleEmbed] });
   }
   if (message.content == "!register") {
-    await registerUser(message);
+    await COMMAND.registerUser(message);
     message.reply("등록까지 1~2분 소요됩니다.");
   }
   if (message.content == "!showAll") {
@@ -210,7 +130,7 @@ client.on("interactionCreate", async (interaction) => {
       interaction.reply(
         `**${interaction.user.username}**님이 '리롤 버튼'을 클릭했습니다.`
       );
-      makeTeam(interaction);
+      COMMAND.makeTeam(interaction);
     } else if (
       interaction.component.data.custom_id === "team1winBtn" &&
       checkDelay
